finalize implementation
