--- packages/dataviews/src/components/dataviews-bulk-actions-toolbar/index.tsx
@@ -1,197 +1,0 @@
-/**
- * WordPress dependencies
- */
-import {
-	ToolbarButton,
-	Toolbar,
-	ToolbarGroup,
-	__unstableMotion as motion,
-	__unstableAnimatePresence as AnimatePresence,
-	__experimentalHStack as HStack,
-} from '@wordpress/components';
-import { useMemo, useState, useRef, useContext } from '@wordpress/element';
-import { _n, sprintf, __ } from '@wordpress/i18n';
-import { closeSmall } from '@wordpress/icons';
-import { useReducedMotion } from '@wordpress/compose';
-import { useRegistry } from '@wordpress/data';
-
-/**
- * Internal dependencies
- */
-import {
-	useSomeItemHasAPossibleBulkAction,
-	BulkSelectionCheckbox,
-} from '../dataviews-bulk-actions';
-import DataViewsContext from '../dataviews-context';
-import { ActionWithModal } from '../dataviews-item-actions';
-import { LAYOUT_GRID, LAYOUT_TABLE } from '../../constants';
-import type { Action } from '../../types';
-import type { ActionTriggerProps } from '../dataviews-item-actions';
-import type { SetSelection } from '../../private-types';
-
-interface ActionButtonProps< Item > {
-	action: Action< Item >;
-	selectedItems: Item[];
-	actionInProgress: string | null;
-	setActionInProgress: ( actionId: string | null ) => void;
-}
-
-interface ToolbarContentProps< Item > {
-	selection: string[];
-	onChangeSelection: SetSelection;
-	data: Item[];
-	actions: Action< Item >[];
-	getItemId: ( item: Item ) => string;
-}
-
-function ActionTrigger< Item >( {
-	action,
-	onClick,
-	isBusy,
-	items,
-}: ActionTriggerProps< Item > ) {
-	const label =
-		typeof action.label === 'string' ? action.label : action.label( items );
-	return (
-		<ToolbarButton
-			disabled={ isBusy }
-			label={ label }
-			icon={ action.icon }
-			isDestructive={ action.isDestructive }
-			size="compact"
-			onClick={ onClick }
-			isBusy={ isBusy }
-			tooltipPosition="top"
-		/>
-	);
-}
-
-const EMPTY_ARRAY: [] = [];
-
-function ActionButton< Item >( {
-	action,
-	selectedItems,
-	actionInProgress,
-	setActionInProgress,
-}: ActionButtonProps< Item > ) {
-	const registry = useRegistry();
-	const selectedEligibleItems = useMemo( () => {
-		return selectedItems.filter( ( item ) => {
-			return ! action.isEligible || action.isEligible( item );
-		} );
-	}, [ action, selectedItems ] );
-	if ( 'RenderModal' in action ) {
-		return (
-			<ActionWithModal
-				key={ action.id }
-				action={ action }
-				items={ selectedEligibleItems }
-				ActionTrigger={ ActionTrigger }
-			/>
-		);
-	}
-	return (
-		<ActionTrigger
-			key={ action.id }
-			action={ action }
-			onClick={ () => {
-				setActionInProgress( action.id );
-				action.callback( selectedItems, {
-					registry,
-				} );
-			} }
-			items={ selectedEligibleItems }
-			isBusy={ actionInProgress === action.id }
-		/>
-	);
-}
-
-function ToolbarContent< Item >( {
-	selection,
-	actions,
-	onChangeSelection,
-	data,
-	getItemId,
-}: ToolbarContentProps< Item > ) {
-	const bulkActions = useMemo(
-		() => actions.filter( ( action ) => action.supportsBulk ),
-		[ actions ]
-	);
-	const selectableItems = useMemo( () => {
-		return data.filter( ( item ) => {
-			return bulkActions.some(
-				( action ) => ! action.isEligible || action.isEligible( item )
-			);
-		} );
-	}, [ data, bulkActions ] );
-
-	const selectedItems = useMemo( () => {
-		return data.filter(
-			( item ) =>
-				selection.includes( getItemId( item ) ) &&
-				selectableItems.includes( item )
-		);
-	}, [ selection, data, getItemId, selectableItems ] );
-	const countToShow =
-		selectedItems.length > 0
-			? selectedItems.length
-			: selectableItems.length;
-	return (
-		<HStack>
-			<BulkSelectionCheckbox
-				selection={ selection }
-				onChangeSelection={ onChangeSelection }
-				data={ data }
-				actions={ actions }
-				getItemId={ getItemId }
-			/>
-			<span>
-				{ sprintf(
-					/* translators: %d: number of items. */
-					_n( '%d item', '%d items', countToShow ),
-					countToShow
-				) }
-			</span>
-		</HStack>
-	);
-}
-
-export default function BulkActionsToolbar() {
-	const {
-		data,
-		selection,
-		actions = EMPTY_ARRAY,
-		onChangeSelection,
-		getItemId,
-	} = useContext( DataViewsContext );
-	const selectedItems = useMemo( () => {
-		return data.filter( ( item ) =>
-			selection.includes( getItemId( item ) )
-		);
-	}, [ selection, data, getItemId ] );
-
-	const actionsToShow = useMemo(
-		() =>
-			actions.filter( ( action ) => {
-				return (
-					action.supportsBulk &&
-					action.icon &&
-					selectedItems.some(
-						( item ) =>
-							! action.isEligible || action.isEligible( item )
-					)
-				);
-			} ),
-		[ actions, selectedItems ]
-	);
-
-	return (
-		<ToolbarContent
-			selection={ selection }
-			onChangeSelection={ onChangeSelection }
-			data={ data }
-			actions={ actions }
-			getItemId={ getItemId }
-		/>
-	);
-}
