--- packages/dataviews/src/components/dataviews-bulk-actions-toolbar/style.scss
@@ -1,5 +1,0 @@
-.dataviews-bulk-actions-toolbar__selection-count {
-	display: flex;
-	align-items: center;
-	margin: 0 $grid-unit-10 0 $grid-unit-10;
-}
