+++ packages/dataviews/src/components/dataviews-bulk-actions/style.scss
@@ -1,7 +1,11 @@
-.dataviews-bulk-actions__modal {
-	z-index: z-index(".dataviews-bulk-actions__modal");
+
+.dataviews-bulk-actions-footer__item-count {
+	color: $gray-900;
+	font-weight: 500;
+	margin-right: $grid-unit-15;
 }
 
-.dataviews-bulk-actions__edit-button.components-button {
-	flex-shrink: 0;
+.dataviews-bulk-actions-footer__container {
+	margin-right: auto;
+	height: $grid-unit-40+$grid-unit-05;
 }
