+++ packages/dataviews/src/components/dataviews-footer/index.tsx
@@ -9,13 +9,30 @@ import { useContext } from '@wordpress/element';
  */
 import DataViewsContext from '../dataviews-context';
 import DataViewsPagination from '../dataviews-pagination';
-import BulkActionsToolbar from '../dataviews-bulk-actions-toolbar';
+import {
+	BulkActionsFooter,
+	useSomeItemHasAPossibleBulkAction,
+} from '../dataviews-bulk-actions';
+import { LAYOUT_GRID, LAYOUT_TABLE } from '../../constants';
+
+const EMPTY_ARRAY: [] = [];
 
 export default function DataViewsFooter() {
 	const {
+		view,
 		paginationInfo: { totalItems = 0, totalPages },
+		data,
+		actions = EMPTY_ARRAY,
 	} = useContext( DataViewsContext );
-	if ( ! totalItems || ! totalPages ) {
+	const hasBulkActions =
+		useSomeItemHasAPossibleBulkAction( actions, data ) &&
+		[ LAYOUT_TABLE, LAYOUT_GRID ].includes( view.type );
+
+	if (
+		! totalItems ||
+		! totalPages ||
+		( totalPages <= 1 && ! hasBulkActions )
+	) {
 		return null;
 	}
 	return (
@@ -26,7 +43,7 @@ export default function DataViewsFooter() {
 				justify="end"
 				className="dataviews-footer"
 			>
-				<BulkActionsToolbar />
+				{ hasBulkActions && <BulkActionsFooter /> }
 				<DataViewsPagination />
 			</HStack>
 		)
