+++ packages/dataviews/src/components/dataviews/index.tsx
@@ -12,7 +12,6 @@ import { useMemo, useState } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import { default as DataViewsBulkActions } from '../dataviews-bulk-actions';
 import DataViewsContext from '../dataviews-context';
 import {
 	default as DataViewsFilters,
